More refactor around layout code.
