+++ public/static/locales/en/common.json
@@ -15,7 +15,7 @@
     "deTitle": "Discovery Environment",
     "discard": "Discard",
     "discovery": "<b>D</b>iscovery",
-    "doi": "DOI",
+    "doiRequests": "DOI Requests",
     "done": "Done",
     "drawerMenuAriaLabel": "drawer menu",
     "edit": "Edit",
